Commit the newest version of x2dhf program.

+++ src/memory.f90
@@ -1,3 +1,3 @@
 module memory
-  integer :: length1,length2,length3,length4,length5,length6,length7,length8
+  integer (KIND=8) :: length1,length2,length3,length4,length5,length6,length7,length8
 end module memory
